--- web/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__, instance_relative_config=True)
-app.config.from_pyfile('config.py')
-app.config.from_pyfile('secrets.py')
